chore: rename button
