+++ my-vue-app/src/components/Header/DropdownMenu.vue
@@ -26,7 +26,7 @@
 
 <script setup>
 import { ref } from 'vue';
-import dropdownArrow from '../../assets/images/icons/check_mark.svg';
+import dropdownArrow from '../../assets/images/icons/check-mark.svg';
 
 const props = defineProps({
   items: {
